+++ src/UglyToad.PdfPig/Filters/RunLengthFilter.cs
@@ -6,7 +6,7 @@
 
     /// <inheritdoc />
     /// <summary>
-    /// The Run Length filterencodes data in a simple byte-oriented format based on run length.
+    /// The Run Length filter encodes data in a simple byte-oriented format based on run length.
     /// The encoded data is a sequence of runs, where each run consists of a length byte followed by 1 to 128 bytes of data.
     /// </summary>
     internal sealed class RunLengthFilter : IFilter
